# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "ApiResponseCorrectionRequestResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CorrectionRequestResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "timestamp"
        ],
        "type": "object"
      },
...
      "ApiResponseListCorrectionRequestResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CorrectionRequestResponse"
            },
            "type": "array"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "timestamp"
        ],
        "type": "object"
      },
...
      "ApiResponseListGuardianStudent": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/GuardianStudent"
            },
            "type": "array"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "timestamp"
        ],
        "type": "object"
      },
...
      "ApiResponseListStudentAttendanceRecord": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/StudentAttendanceRecord"
            },
            "type": "array"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "timestamp"
        ],
        "type": "object"
      },
...
      "ApiResponsePageResponseCorrectionRequestResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PageResponseCorrectionRequestResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "timestamp"
        ],
        "type": "object"
      },
...
      "ApiResponseSectionAttendanceView": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SectionAttendanceView"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "timestamp"
        ],
        "type": "object"
      },
...
      "ApiResponseStudentDetail": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/StudentDetail"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "timestamp"
        ],
        "type": "object"
      },
...
      "AttendanceEntryRequest": {
        "properties": {
          "remarks": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "status": {
            "enum": [
              "PRESENT",
              "ABSENT",
              "LATE",
              "HALF_DAY",
              "EXCUSED_LEAVE",
              "HOLIDAY"
            ],
            "type": "string"
          },
          "studentId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "status",
          "studentId"
        ],
        "type": "object"
      },
      "AttendanceStudentMark": {
        "properties": {
          "admissionNumber": {
            "type": "string"
          },
          "editable": {
            "type": "boolean"
          },
          "fullName": {
            "type": "string"
          },
          "markId": {
            "format": "uuid",
            "type": "string"
          },
          "remarks": {
            "type": "string"
          },
          "rollNumber": {
            "type": "string"
          },
          "status": {
            "enum": [
              "PRESENT",
              "ABSENT",
              "LATE",
              "HALF_DAY",
              "EXCUSED_LEAVE",
              "HOLIDAY"
            ],
            "type": "string"
          },
          "studentId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "admissionNumber",
          "editable",
          "fullName",
          "studentId"
        ],
        "type": "object"
      },
...
      "ComplianceSummary": {
        "properties": {
          "apaarConsentGiven": {
            "type": "boolean"
          },
          "apaarConsentGivenAt": {
            "format": "date-time",
            "type": "string"
          },
          "apaarConsentGivenBy": {
            "type": "string"
          },
          "boardRegistrationNumber": {
            "type": "string"
          },
          "hasApaarId": {
            "type": "boolean"
          },
          "hasCasteCategory": {
            "type": "boolean"
          },
          "hasReligion": {
            "type": "boolean"
          },
          "hasSpecialCategory": {
            "type": "boolean"
          },
          "penUdiseId": {
            "type": "string"
          }
        },
        "required": [
          "apaarConsentGiven",
          "hasApaarId",
          "hasCasteCategory",
          "hasReligion",
          "hasSpecialCategory"
        ],
        "type": "object"
      },
      "ContactDetail": {
        "properties": {
          "address": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "CorrectionRequestResponse": {
        "properties": {
          "attendanceDate": {
            "format": "date",
            "type": "string"
          },
          "attendanceMarkId": {
            "format": "uuid",
            "type": "string"
          },
          "decidedAt": {
            "format": "date-time",
            "type": "string"
          },
          "decision": {
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ],
            "type": "string"
          },
          "decisionNote": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "previousStatus": {
            "enum": [
              "PRESENT",
              "ABSENT",
              "LATE",
              "HALF_DAY",
              "EXCUSED_LEAVE",
              "HOLIDAY"
            ],
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "requestedAt": {
            "format": "date-time",
            "type": "string"
          },
          "requestedStatus": {
            "enum": [
              "PRESENT",
              "ABSENT",
              "LATE",
              "HALF_DAY",
              "EXCUSED_LEAVE",
              "HOLIDAY"
            ],
            "type": "string"
          },
          "studentId": {
            "format": "uuid",
            "type": "string"
          },
          "studentName": {
            "type": "string"
          }
        },
        "required": [
          "attendanceDate",
          "attendanceMarkId",
          "decision",
          "id",
          "previousStatus",
          "reason",
          "requestedAt",
          "requestedStatus",
          "studentId",
          "studentName"
        ],
        "type": "object"
      },
...
      "CurrentEnrolment": {
        "properties": {
          "className": {
            "type": "string"
          },
          "rollNumber": {
            "type": "string"
          },
          "sectionName": {
            "type": "string"
          },
          "sessionName": {
            "type": "string"
          }
        },
        "required": [
          "className",
          "sectionName",
          "sessionName"
        ],
        "type": "object"
      },
...
      "DecideCorrectionRequest": {
        "properties": {
          "decision": {
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ],
            "type": "string"
          },
          "note": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "decision"
        ],
        "type": "object"
      },
...
      "Enrolment": {
        "properties": {
          "active": {
            "type": "boolean"
          },
          "classId": {
            "format": "uuid",
            "type": "string"
          },
          "className": {
            "type": "string"
          },
          "enrolledOn": {
            "format": "date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "rollNumber": {
            "type": "string"
          },
          "sectionId": {
            "format": "uuid",
            "type": "string"
          },
          "sectionName": {
            "type": "string"
          },
          "sessionId": {
            "format": "uuid",
            "type": "string"
          },
          "sessionName": {
            "type": "string"
          }
        },
        "required": [
          "active",
          "enrolledOn",
          "id",
          "sectionId",
          "sessionId"
        ],
        "type": "object"
      },
...
      "GuardianStudent": {
        "properties": {
          "admissionNumber": {
            "type": "string"
          },
          "currentEnrolment": {
            "$ref": "#/components/schemas/CurrentEnrolment"
          },
          "fullName": {
            "type": "string"
          },
          "primary": {
            "type": "boolean"
          },
          "relation": {
            "enum": [
              "FATHER",
              "MOTHER",
              "GUARDIAN",
              "LOCAL_GUARDIAN",
              "OTHER"
            ],
            "type": "string"
          },
          "studentId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "admissionNumber",
          "fullName",
          "primary",
          "relation",
          "studentId"
        ],
        "type": "object"
      },
...
      "MarkAttendanceRequest": {
        "properties": {
          "attendanceDate": {
            "format": "date",
            "type": "string"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/AttendanceEntryRequest"
            },
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "attendanceDate",
          "entries"
        ],
        "type": "object"
      },
...
      "PageResponseCorrectionRequestResponse": {
        "properties": {
          "content": {
            "items": {
              "$ref": "#/components/schemas/CorrectionRequestResponse"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "totalElements": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "content",
          "page",
          "size",
          "totalElements",
          "totalPages"
        ],
        "type": "object"
      },
...
      "Pageable": {
        "properties": {
          "page": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          },
          "sort": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "PreviousSchoolDetail": {
        "properties": {
          "previousSchoolBoard": {
            "type": "string"
          },
          "previousSchoolName": {
            "type": "string"
          },
          "reasonForLeaving": {
            "type": "string"
          },
          "transferCertificateIssuedOn": {
            "format": "date",
            "type": "string"
          },
          "transferCertificateNumber": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "RequestCorrectionRequest": {
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "requestedStatus": {
            "enum": [
              "PRESENT",
              "ABSENT",
              "LATE",
              "HALF_DAY",
              "EXCUSED_LEAVE",
              "HOLIDAY"
            ],
            "type": "string"
          }
        },
        "required": [
          "reason",
          "requestedStatus"
        ],
        "type": "object"
      },
...
      "SaveStudentRequest": {
        "properties": {
          "admissionNumber": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "admittedOn": {
            "format": "date",
            "type": "string"
          },
          "dateOfBirth": {
            "format": "date",
            "type": "string"
          },
          "fullName": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "gender": {
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "TRANSFERRED",
              "GRADUATED",
              "WITHDRAWN"
            ],
            "type": "string"
          }
        },
        "required": [
          "admissionNumber",
          "dateOfBirth",
          "fullName",
          "gender",
          "status"
        ],
        "type": "object"
      },
...
      "SectionAttendanceView": {
        "properties": {
          "academicSessionId": {
            "format": "uuid",
            "type": "string"
          },
          "attendanceDate": {
            "format": "date",
            "type": "string"
          },
          "className": {
            "type": "string"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/AttendanceStudentMark"
            },
            "type": "array"
          },
          "locked": {
            "type": "boolean"
          },
          "sectionId": {
            "format": "uuid",
            "type": "string"
          },
          "sectionName": {
            "type": "string"
          }
        },
        "required": [
          "academicSessionId",
          "attendanceDate",
          "className",
          "entries",
          "locked",
          "sectionId",
          "sectionName"
        ],
        "type": "object"
      },
...
      "StudentAttendanceRecord": {
        "properties": {
          "attendanceDate": {
            "format": "date",
            "type": "string"
          },
          "editable": {
            "type": "boolean"
          },
          "markId": {
            "format": "uuid",
            "type": "string"
          },
          "remarks": {
            "type": "string"
          },
          "status": {
            "enum": [
              "PRESENT",
              "ABSENT",
              "LATE",
              "HALF_DAY",
              "EXCUSED_LEAVE",
              "HOLIDAY"
            ],
            "type": "string"
          }
        },
        "required": [
          "attendanceDate",
          "editable",
          "markId",
          "status"
        ],
        "type": "object"
      },
      "StudentDetail": {
        "properties": {
          "admissionNumber": {
            "type": "string"
          },
          "admittedOn": {
            "format": "date",
            "type": "string"
          },
          "compliance": {
            "$ref": "#/components/schemas/ComplianceSummary"
          },
          "contact": {
            "$ref": "#/components/schemas/ContactDetail"
          },
          "currentEnrolment": {
            "$ref": "#/components/schemas/CurrentEnrolment"
          },
          "dateOfBirth": {
            "format": "date",
            "type": "string"
          },
          "enrolments": {
            "items": {
              "$ref": "#/components/schemas/Enrolment"
            },
            "type": "array"
          },
          "fullName": {
            "type": "string"
          },
          "gender": {
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ],
            "type": "string"
          },
          "guardians": {
            "items": {
              "$ref": "#/components/schemas/StudentGuardian"
            },
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "medical": {
            "$ref": "#/components/schemas/MedicalSummary"
          },
          "previousSchool": {
            "$ref": "#/components/schemas/PreviousSchoolDetail"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "TRANSFERRED",
              "GRADUATED",
              "WITHDRAWN"
            ],
            "type": "string"
          }
        },
        "required": [
          "admissionNumber",
          "compliance",
          "dateOfBirth",
          "enrolments",
          "fullName",
          "gender",
          "guardians",
          "id",
          "medical",
          "status"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "School management system for Indian K-12 schools",
    "license": {
      "name": "Proprietary"
    },
    "title": "Chalkbase API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/attendance/correction-requests": {
      "get": {
        "operationId": "pending",
        "parameters": [
          {
            "in": "query",
            "name": "pageable",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseCorrectionRequestResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-correction-controller"
        ]
      }
    },
    "/api/attendance/correction-requests/{id}/decision": {
      "post": {
        "operationId": "decide",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideCorrectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCorrectionRequestResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-correction-controller"
        ]
      }
    },
    "/api/attendance/marks/{markId}/correction-requests": {
      "get": {
        "operationId": "correctionHistory",
        "parameters": [
          {
            "in": "path",
            "name": "markId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListCorrectionRequestResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-controller"
        ]
      },
      "post": {
        "operationId": "requestCorrection",
        "parameters": [
          {
            "in": "path",
            "name": "markId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCorrectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCorrectionRequestResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-controller"
        ]
      }
    },
    "/api/attendance/sections/{sectionId}": {
      "get": {
        "operationId": "view",
        "parameters": [
          {
            "in": "path",
            "name": "sectionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSectionAttendanceView"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-controller"
        ]
      },
      "post": {
        "operationId": "mark",
        "parameters": [
          {
            "in": "path",
            "name": "sectionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkAttendanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSectionAttendanceView"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-controller"
        ]
      }
    },
    "/api/attendance/students/{studentId}": {
      "get": {
        "operationId": "studentHistory",
        "parameters": [
          {
            "in": "path",
            "name": "studentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListStudentAttendanceRecord"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-controller"
        ]
      }
    },
...
    "/api/guardians/{id}/students": {
      "get": {
        "operationId": "students",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListGuardianStudent"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "guardian-controller"
        ]
      }
    },
...
    "/api/students/{id}": {
      "get": {
        "operationId": "find",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudentDetail"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "student-controller"
        ]
      },
      "put": {
        "operationId": "update_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveStudentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudentDetail"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "student-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost"
    }
  ]
}